+++ storybook/src/_common/renderComponentVariantTypes.ts
@@ -3,8 +3,8 @@
  * Copyright Gemeente Amsterdam
  */
 
+import type { IconProps } from '@amsterdam/design-system-react'
 import type { Meta } from '@storybook/react-vite'
-import type { JSX, SVGProps } from 'react'
 
 export type renderComponentVariantsParams = {
   args: Meta['args'] // Storybook args applied to every instance
@@ -32,7 +32,7 @@ export type DocgenInfo = {
 export type PropWithValues = {
   name: string
   propType: string | undefined
-  values: string[] | number[] | boolean[] | ((props: SVGProps<SVGSVGElement>) => JSX.Element)[]
+  values: string[] | number[] | boolean[] | IconProps['svg'][]
 }
 
 export type CompletePropsWithValues = (
@@ -43,7 +43,7 @@ export type CompletePropsWithValues = (
     }
   | {
       hasIcon: {
-        icon: (props: SVGProps<SVGSVGElement>) => JSX.Element
+        icon: IconProps['svg']
       }
       name: string
       values: PropWithValues['values']
@@ -52,11 +52,11 @@ export type CompletePropsWithValues = (
 
 export type BuildComponentPropsParams = {
   args: Meta['args']
-  hasIcon?: { icon: (props: SVGProps<SVGSVGElement>) => JSX.Element } | null
+  hasIcon?: { icon: IconProps['svg'] } | null
   propName: string
   size?: string | undefined
   sizePropName: string | string[]
   state: string
   values?: PropWithValues['values']
-  variant: string | number | boolean | ((props: SVGProps<SVGSVGElement>) => JSX.Element)
+  variant: string | number | boolean | IconProps['svg']
 }