+++ storybook/src/pages/internal/TablePage/common/index.ts
@@ -1,0 +1,10 @@
+/**
+ * @license EUPL-1.2+
+ * Copyright Gemeente Amsterdam
+ */
+
+export * from './AddressTableBody'
+export * from './AddressTableHeaderRow'
+export * from './bagAddresses'
+export * from './options'
+export * from './sortAddresses'